升级springboot3ing

+++ src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -11,10 +11,11 @@ import com.hmdp.service.IFollowService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.hmdp.service.IUserService;
 import com.hmdp.utils.UserHolder;
+import jakarta.annotation.Resource;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -80,7 +81,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     public Result isFollow(Long followUserId) {
         Long userId = UserHolder.getUser().getId();
         //select count(*) from tb_follow where user_id=? and follow_user_id = ?
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
+        Long count = query().eq("user_id", userId).eq("follow_user_id", followUserId).count();
 
         return Result.ok(count>0);
